flask-app/requirements.txt
flask==2.3.3
mysql-connector-python==9.2.0
